Enable art/tools/ahat to use libjavacrypto.
`ahat-tests` requires the `libjavacrypto.so` at runtime. This change
allows ahat-tests to use them.

This is the `prebuilt` version of `libjavacrypto`, but it takes a long
time for the source code version to be officially pushed to the
prebuilt version. Since adding visibility is safe in itself, we will
update the visibility of `libjavacrypto` together with the source code
version for now.

Furthermore, as the AOSP version is currently an older version,
submitting directly to AOSP will cause a merge conflict. Therefore, it
needs to be merged into the main branch first.

Bug: 361491302
Bug: 342330305
Test: m -j libjavacrypto; atest ahat-tests
Change-Id: I16e7ee52222220e896f90b702d5bfe3b835794cc