Enable art/tools/ahat to use libjavacrypto.
`ahat-tests` requires the `libjavacrypto.so` at runtime. This change
allows ahat-tests to use them.

This is the `prebuilt` version of `libjavacrypto`. The `source code`
version has already been merged in `aosp/3235303`. This change is
necessary because, although `ahat-tests` are on `AOSP` and normally
use the `source code` version, the `presubmit tests` on the
`aosp-master-art` branch use the `prebuilt` version. Currently, the
code on git-master and aosp has diverged. Manual cherry-picking is
needed on aosp-main and aosp-main-future to resolve this merge
conflict.

This change was previously merged in `ag/28947702` but was overwritten
in `ag/28995483`, hence resubmitting.

Bug: 361491302
Test: m -j libjavacrypto; atest ahat-tests
Change-Id: Id44d7d35e9f266230979d695d373970d99acea3a
Merged-In: Id44d7d35e9f266230979d695d373970d99acea3a
(cherry picked from commit 80e685e21ed1384433864ea01cc6e918d3a1b3c6)